SDLC Iter 1: Move MapNode to core-model, fix circular dependency
MapNode was defined in feature-map but imported by core-data
(MapNodeLoader, MultiActMapNodeLoader). This created a circular
dependency since core-data can't depend on feature modules.

Fix:
- Create core-model/MapNode.kt with the data class
- Remove MapNode data class from feature-map/MapViewModel.kt
- Update MapNodeLoader import: com.chimera.model.MapNode
- Update MultiActMapNodeLoader import: com.chimera.model.MapNode
- Add MapNode import to feature-map/MapScreen.kt
- Add MapNode import to feature-map/MapViewModel.kt

DuelEngine import (originally planned): already fixed by Copilot
in sprint v1.8.0 -- upgraded to CombatEngine in chimera-core with
correct import in DuelScreen/DuelViewModel.
